requirements.txt
playwright==1.40.0
pandas==2.1.4
openpyxl==3.1.2
requests==2.31.0
loguru==0.7.2
pyyaml==6.0.1 